Update rake task to use JSON.pretty_generate so output is always in JSON
Now that the config accepts multiple formats, the output of Config.read
was switched from a JSON string to a hash. Inorder to always return
a json string from the rake task pretty_generate is used. This maintains
a degree of backwards compatbility incase someone built an API around
that task's output.